+++ test/clj_kondo/main_test.clj
@@ -1748,14 +1748,15 @@ foo/foo ;; this does use the private var
     (is (empty? (lint! "(comment (def x 1) (def x 2))")))
     (is (empty? (lint! "(comment (def x 1)) (def x 2)")))))
 
-(deftest unreachable-code-test
+(deftest constant-condition-test
   (assert-submaps
    '({:file "<stdin>",
       :row 1,
       :col 15,
       :level :warning,
-      :message "unreachable code"})
-   (lint! "(cond :else 1 (odd? 1) 2)")))
+      :message "Unreachable code"})
+   (lint! "(cond :else 1 (odd? 1) 2)"
+          {:linters {:constant-condition {:level :warning}}})))
 
 (deftest dont-crash-analyzer-test
   (doseq [example ["(let)" "(if-let)" "(when-let)" "(loop)" "(doseq)"]